test-requirements.txt
coverage>=3.6
discover
dpath>=1.3.2
fixtures>=0.3.14
# Hacking already pins down pep8, pyflakes and flake8
hacking>=0.10.0,<0.11
mock>=1.0
requests>=1.2.1,!=2.4.0
requests-mock>=0.4.0  # Apache-2.0
salt
testrepository>=0.0.18
testscenarios>=0.4
testtools>=0.9.34
yamllint>=1.2.0
os-apply-config

python-heatclient>=1.2.0
python-keystoneclient>=0.10.0
python-openstackclient>=2.1.0
